Scss Webpack.config, Changed BookingCalendar class name (Calendar), and created a view class (BookingCalendar)
