+++ package.json
@@ -21,8 +21,11 @@
   "devDependencies": {
     "@types/jest": "^26.0.22",
     "@types/node": "^14.14.37",
+    "autoprefixer": "^9.0.0",
     "css-loader": "^5.2.4",
+    "cssnano": "^4.0.5",
     "express": "^4.17.1",
+    "html-webpack-plugin": "^5.3.1",
     "jest": "^26.6.3",
     "jest-cli": "^26.6.3",
     "mini-css-extract-plugin": "^1.5.0",
@@ -36,8 +39,6 @@
     "tsconfig-paths": "^3.9.0",
     "typescript": "^4.2.4",
     "webpack": "^5.35.0",
-    "webpack-cli": "^4.6.0",
-    "autoprefixer": "^9.0.0",
-    "cssnano": "^4.0.5"
+    "webpack-cli": "^4.6.0"
   }
 }
