--- src/assets/sass/_colours.sass
@@ -1,1 +1,0 @@
-$
